Extract PriceTextField component to common folder
